Update whatsapp template view
Instead of using the default 'message' argument, we now need to account for other, additional parameters.

These can be passed to the endpoint by numbering your param arguments. e.g. `0=R25&1=24` will pass the parameters `R25` and `24` to the template, in that order.

Note that this endpoint only orders the arguments, it does not attempt to skip over arguments. Thus `0=R25&1=24`, `1=R25&2=24` and `1=R25&99=24` will all be treated in an identical manner.